Make it so all posts are stored in one central object
- Changes the way we do timelines (all posts are stored in a
central, id-addressed object, and timelines are now just arrays
of ids)
- Adapts existing components to this new way of doing things

This change is being made to make it easier to change stuff
about posts later down the line; i.e. when a post is favorited,
we now only have to change one object, rather than having to
iterate over all timelines to find and modify all possible
instances.

I'm not sure I like this approach just yet, so consider this an
experiment for now.